Create README.md
Here's a preview of how the modified trivia game will work and what it will show:

When you open the trivia game, you will see a question displayed on the screen along with a list of answer options. You can read the question and select your answer by clicking on one of the options. The selected answer will be highlighted.

In the background, a pleasant background music will start playing as soon as the game begins. The music will create a lively and engaging atmosphere while you navigate through the trivia questions.

Once you have selected an answer, you can click the "Submit" button to submit your answer. If your answer is correct, your score will increase by one point. The game will then automatically move on to the next question, and the updated score will be displayed.

If you want to go back and change your answer for the current question, you can click the "Previous" button to navigate to the previous question. The background music will continue playing as you navigate through the questions.

After you have answered all the questions, a congratulatory message will pop up, indicating that you have finished the trivia. The message will also display your final score out of the total number of questions. At this point, you can choose to play the game again by clicking the "Next" button, which will restart the game, reset the score, and randomly select a new question.

In summary, the modified trivia game will feature background music that plays throughout the game, adding an enjoyable audio element to the trivia experience. It will enhance the overall ambiance and make the game more immersive and entertaining for the players.